requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
djangorestframework==3.10.2
pytz==2018.7